image_des_onnx/requirements.txt
torch
transformers
onnx
onnxruntime
Pillow
